Make @refresh_cache decorator take update_cells argument
A recent change modifies the @refresh_cache decorator so that it only
updates instance cache if needed. For the get_instance_nw_info()
method to take advantage of this, it must use the decorator instead of
calling update_instance_cache_with_nw_info() directly. To do this, the
decorator must be modified to take the update_cells argument.

Related blueprint: smarter-network-cache-update

Change-Id: Id56fcb87458a88241f3929de7312da18b9cf5b9f